server/requirements.txt
eventlet==0.31.1
Flask==2.0.1
Flask_Cors==3.0.10
Flask_JWT_Extended==4.4.4
Flask_SocketIO==5.1.0
Flask_SQLAlchemy==2.5.1
PyJWT==2.5.0
python-dotenv==0.21.0
SQLAlchemy_Utils==0.38.2
Werkzeug==2.0.1